purely sequential dw for better testng

+++ send_control.cpp
@@ -178,7 +178,7 @@ tint    Channel::CwndRateNextSendTime () {
         return data_out_.front().time + ack_timeout();
     }*/
     // start test
-    if (data_out_size_<(int)cwnd_ || (int)cwnd_ > 1) {
+    if (data_out_size_<(int)cwnd_ || cwnd_ > 1) {
         dprintf("%s #%" PRIu32 " sendctrl send interval %" PRIi64 "us (cwnd %.2f, data_out %" PRIu32 ")\n",
                 tintstr(),id_,send_interval_,cwnd_,data_out_size_);
         return last_data_out_time_ + send_interval_ - timer_delay_;
